Adds x86 feature guards
