+++ algorithms/Cargo.toml
@@ -82,14 +82,10 @@ name = "groth16"
 path = "examples/snark/groth16.rs"
 required-features = ["snark"]
 
-[dependencies]
-voracious_radix_sort = { version = "1.1.0" }
-
 [dependencies.snarkvm-curves]
 path = "../curves"
 version = "0.7.5"
 default-features = false
-features = ["prefetch"]
 
 [dependencies.snarkvm-fields]
 path = "../fields"
