+++ algorithms/src/msm/variable_base/mod.rs
@@ -63,8 +63,7 @@ impl VariableBaseMSM {
             }
         }
         // standard::msm_standard(bases, scalars)
-        standard::msm_standard_batched(bases, scalars, 0) // Option A
-        // standard::msm_standard_batched(bases, scalars, 1) // Option B
+        standard::msm_standard_batched(bases, scalars)
     }
 }
