+++ algorithms/src/msm/variable_base/standard.rs
@@ -15,7 +15,7 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use snarkvm_curves::{traits::AffineCurve, Group, ProjectiveCurve};
-use snarkvm_fields::{FieldParameters, One, PrimeField, Zero};
+use snarkvm_fields::{One, PrimeField, Zero};
 use snarkvm_utilities::{cfg_into_iter, BigInteger};
 
 #[cfg(feature = "parallel")]
@@ -86,7 +86,7 @@ pub(super) fn msm_standard<G: AffineCurve>(
         false => crate::msm::ln_without_floats(scalars.len()) + 2,
     };
 
-    let num_bits = <G::ScalarField as PrimeField>::Parameters::MODULUS_BITS as usize;
+    let num_bits = <G::ScalarField as PrimeField>::size_in_bits();
 
     // Each window is of size `c`.
     // We divide up the bits 0..num_bits into windows of size `c`, and
@@ -110,7 +110,6 @@ pub(super) fn msm_standard<G: AffineCurve>(
 pub(super) fn msm_standard_batched<G: AffineCurve>(
     bases: &[G],
     scalars: &[<G::ScalarField as PrimeField>::BigInteger],
-    option: usize,
 ) -> G::Projective {
     let c = match scalars.len() < 32 {
         true => 1,
@@ -119,17 +118,12 @@ pub(super) fn msm_standard_batched<G: AffineCurve>(
 
     let num_bits = <G::ScalarField as PrimeField>::size_in_bits();
     let zero = G::Projective::zero();
-    let window_starts: Vec<_> = (0..num_bits).step_by(c).collect();
-
-    #[cfg(feature = "parallel")]
-    let window_starts_iter = window_starts.into_par_iter();
-    #[cfg(not(feature = "parallel"))]
-    let window_starts_iter = window_starts.into_iter();
 
     // Each window is of size `c`.
     // We divide up the bits 0..num_bits into windows of size `c`, and
     // in parallel process each such window.
-    let window_sums: Vec<_> = window_starts_iter
+    let window_sums: Vec<_> = cfg_into_iter!(0..num_bits)
+        .step_by(c)
         .map(|w_start| {
             // We don't need the "zero" bucket, so we only have 2^c - 1 buckets
             let log2_n_bucket = if (w_start % c) != 0 { w_start % c } else { c };
@@ -151,11 +145,7 @@ pub(super) fn msm_standard_batched<G: AffineCurve>(
                 })
                 .collect();
 
-            let buckets = match option {
-                0 => batch_bucketed_add_option_a::<G>(n_buckets, &bases[..], &mut bucket_positions[..]),
-                1 => batch_bucketed_add_option_b::<G>(n_buckets, &bases[..], &mut bucket_positions[..]),
-                _ => unreachable!(),
-            };
+            let buckets = batch_add_bucket::<G>(n_buckets, &bases[..], &mut bucket_positions[..]);
 
             let mut res = zero;
             let mut running_sum = G::Projective::zero();
@@ -191,7 +181,6 @@ pub(super) fn msm_standard_batched<G: AffineCurve>(
 pub const BATCH_SIZE: usize = 4096;
 
 use core::cmp::Ordering;
-use voracious_radix_sort::*;
 
 #[cfg(not(feature = "std"))]
 use crate::fft::domain::log2;
@@ -208,166 +197,10 @@ impl PartialOrd for BucketPosition {
     }
 }
 
-impl Radixable<u32> for BucketPosition {
-    type Key = u32;
-
-    #[inline]
-    fn key(&self) -> Self::Key {
-        self.bucket
-    }
-}
-
 impl PartialEq for BucketPosition {
     fn eq(&self, other: &Self) -> bool {
         self.bucket == other.bucket
     }
-}
-
-#[inline]
-pub fn batch_bucketed_add_option_a<C: AffineCurve>(
-    buckets: usize,
-    elems: &[C],
-    bucket_positions: &mut [BucketPosition],
-) -> Vec<C> {
-    // assert_eq!(elems.len(), bucket_positions.len());
-    assert!(elems.len() > 0);
-
-    dlsd_radixsort(bucket_positions, 8);
-
-    let mut len = bucket_positions.len();
-    let mut all_ones = true;
-    let mut new_len = 0; // len counter
-    let mut glob = 0; // global counters
-    let mut loc = 1; // local counter
-    let mut batch = 0; // batch counter
-    let mut instr = Vec::<(u32, u32)>::with_capacity(BATCH_SIZE);
-    let mut new_elems = Vec::<C>::with_capacity(elems.len() * 3 / 8);
-
-    let mut scratch_space = Vec::<Option<C>>::with_capacity(BATCH_SIZE / 2);
-
-    // In the first loop, we copy the results of the first in place addition tree
-    // to a local vector, new_elems
-    // Subsequently, we perform all the operations in place
-    while glob < len {
-        let current_bucket = bucket_positions[glob].bucket;
-        while glob + 1 < len && bucket_positions[glob + 1].bucket == current_bucket {
-            glob += 1;
-            loc += 1;
-        }
-        if current_bucket >= buckets as u32 {
-            loc = 1;
-        } else if loc > 1 {
-            // all ones is false if next len is not 1
-            if loc > 2 {
-                all_ones = false;
-            }
-            let is_odd = loc % 2 == 1;
-            let half = loc / 2;
-            for i in 0..half {
-                instr.push((
-                    bucket_positions[glob - (loc - 1) + 2 * i].position,
-                    bucket_positions[glob - (loc - 1) + 2 * i + 1].position,
-                ));
-                bucket_positions[new_len + i] =
-                    BucketPosition { bucket: current_bucket, position: (new_len + i) as u32 };
-            }
-            if is_odd {
-                instr.push((bucket_positions[glob].position, !0u32));
-                bucket_positions[new_len + half] =
-                    BucketPosition { bucket: current_bucket, position: (new_len + half) as u32 };
-            }
-            // Reset the local_counter and update state
-            new_len += half + (loc % 2);
-            batch += half;
-            loc = 1;
-
-            if batch >= BATCH_SIZE / 2 {
-                // We need instructions for copying data in the case
-                // of noops. We encode noops/copies as !0u32
-                C::batch_add_write(&elems[..], &instr[..], &mut new_elems, &mut scratch_space);
-
-                instr.clear();
-                batch = 0;
-            }
-        } else {
-            instr.push((bucket_positions[glob].position, !0u32));
-            bucket_positions[new_len] = BucketPosition { bucket: current_bucket, position: new_len as u32 };
-            new_len += 1;
-        }
-        glob += 1;
-    }
-    if instr.len() > 0 {
-        C::batch_add_write(&elems[..], &instr[..], &mut new_elems, &mut scratch_space);
-        instr.clear();
-    }
-    glob = 0;
-    batch = 0;
-    loc = 1;
-    len = new_len;
-    new_len = 0;
-
-    while !all_ones {
-        all_ones = true;
-        while glob < len {
-            let current_bucket = bucket_positions[glob].bucket;
-            while glob + 1 < len && bucket_positions[glob + 1].bucket == current_bucket {
-                glob += 1;
-                loc += 1;
-            }
-            if current_bucket >= buckets as u32 {
-                loc = 1;
-            } else if loc > 1 {
-                // all ones is false if next len is not 1
-                if loc != 2 {
-                    all_ones = false;
-                }
-                let is_odd = loc % 2 == 1;
-                let half = loc / 2;
-                for i in 0..half {
-                    instr.push((
-                        bucket_positions[glob - (loc - 1) + 2 * i].position,
-                        bucket_positions[glob - (loc - 1) + 2 * i + 1].position,
-                    ));
-                    bucket_positions[new_len + i] = bucket_positions[glob - (loc - 1) + 2 * i];
-                }
-                if is_odd {
-                    bucket_positions[new_len + half] = bucket_positions[glob];
-                }
-                // Reset the local_counter and update state
-                new_len += half + (loc % 2);
-                batch += half;
-                loc = 1;
-
-                if batch >= BATCH_SIZE / 2 {
-                    C::batch_add_in_place_same_slice(&mut new_elems[..], &instr[..]);
-                    instr.clear();
-                    batch = 0;
-                }
-            } else {
-                bucket_positions[new_len] = bucket_positions[glob];
-                new_len += 1;
-            }
-            glob += 1;
-        }
-        if instr.len() > 0 {
-            C::batch_add_in_place_same_slice(&mut new_elems[..], &instr[..]);
-            instr.clear();
-        }
-        glob = 0;
-        batch = 0;
-        loc = 1;
-        len = new_len;
-        new_len = 0;
-    }
-
-    let zero = C::zero();
-    let mut res = vec![zero; buckets];
-
-    for i in 0..len {
-        let (pos, buc) = (bucket_positions[i].position, bucket_positions[i].bucket);
-        res[buc as usize] = new_elems[pos as usize];
-    }
-    res
 }
 
 /// Chunks vectorised instructions into a size that does not require
@@ -389,11 +222,7 @@ fn get_chunked_instr<T: Clone>(instr: &[T], batch_size: usize) -> Vec<Vec<T>> {
     res
 }
 
-pub fn batch_bucketed_add_option_b<C: AffineCurve>(
-    buckets: usize,
-    elems: &[C],
-    bucket_assign: &[BucketPosition],
-) -> Vec<C> {
+pub fn batch_add_bucket<C: AffineCurve>(buckets: usize, elems: &[C], bucket_assign: &[BucketPosition]) -> Vec<C> {
     let mut elems = elems.to_vec();
     let num_split = 2i32.pow(log2(buckets) / 2 + 2) as usize;
     let split_size = (buckets - 1) / num_split + 1;
@@ -451,8 +280,7 @@ pub fn batch_bucketed_add_option_b<C: AffineCurve>(
         }
     }
 
-    let zero = C::zero();
-    let mut res = vec![zero; buckets];
+    let mut res = vec![C::zero(); buckets];
 
     for (i, to_add) in index.iter().enumerate() {
         if to_add.len() == 1 {
