+++ curves/src/templates/twisted_edwards_extended/affine.rs
@@ -16,6 +16,7 @@
 
 use crate::{
     impl_edwards_curve_serializer,
+    prefetch_slice,
     templates::twisted_edwards_extended::Projective,
     traits::{AffineCurve, Group, MontgomeryParameters, ProjectiveCurve, TwistedEdwardsParameters as Parameters},
 };
@@ -261,8 +262,18 @@ impl<P: Parameters> AffineCurve for Affine<P> {
     // Total cost: 12 mul. Projective formulas: 11 mul.
     fn batch_add_in_place_same_slice(bases: &mut [Self], index: &[(u32, u32)]) {
         let mut inversion_tmp = P::BaseField::one();
+
+        #[cfg(target_arch = "x86_64")]
+        let mut prefetch_iter = index.iter();
+        #[cfg(target_arch = "x86_64")]
+        {
+            prefetch_iter.next();
+        }
+
         // We run two loops over the data separated by an inversion
         for (idx, idy) in index.iter() {
+            #[cfg(target_arch = "x86_64")]
+            prefetch_slice!(bases, bases, prefetch_iter);
             let (mut a, mut b) = if idx < idy {
                 let (x, y) = bases.split_at_mut(*idy as usize);
                 (&mut x[*idx as usize], &mut y[0])
@@ -275,8 +286,17 @@ impl<P: Parameters> AffineCurve for Affine<P> {
 
         inversion_tmp = inversion_tmp.inverse().unwrap(); // this is always in Fp*
 
+        #[cfg(target_arch = "x86_64")]
+        let mut prefetch_iter = index.iter().rev();
+        #[cfg(target_arch = "x86_64")]
+        {
+            prefetch_iter.next();
+        }
+
         for (idx, idy) in index.iter().rev() {
-            let (a, b) = if idx < idy {
+            #[cfg(target_arch = "x86_64")]
+            prefetch_slice!(bases, bases, prefetch_iter);
+            let (mut a, b) = if idx < idy {
                 let (x, y) = bases.split_at_mut(*idy as usize);
                 (&mut x[*idx as usize], y[0])
             } else {
@@ -285,40 +305,6 @@ impl<P: Parameters> AffineCurve for Affine<P> {
             };
             batch_add_loop_2!(a, b, inversion_tmp);
         }
-    }
-
-    fn batch_add_write(
-        lookup: &[Self],
-        index: &[(u32, u32)],
-        new_elems: &mut Vec<Self>,
-        scratch_space: &mut Vec<Option<Self>>,
-    ) {
-        let mut inversion_tmp = P::BaseField::one();
-
-        for (idx, idy) in index.iter() {
-            if *idy == !0u32 {
-                new_elems.push(lookup[*idx as usize]);
-                scratch_space.push(None);
-            } else {
-                let (mut a, mut b) = (lookup[*idx as usize], lookup[*idy as usize]);
-                batch_add_loop_1!(a, b, inversion_tmp);
-                new_elems.push(a);
-                scratch_space.push(Some(b));
-            }
-        }
-
-        inversion_tmp = inversion_tmp.inverse().unwrap(); // this is always in Fp*
-
-        for (a, op_b) in new_elems.iter_mut().rev().zip(scratch_space.iter().rev()) {
-            match op_b {
-                Some(b) => {
-                    let b_ = *b;
-                    batch_add_loop_2!(a, b_, inversion_tmp);
-                }
-                None => (),
-            };
-        }
-        scratch_space.clear();
     }
 }
